+++ example/lib/showcase_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:iron_widgets/iron_widgets.dart';
 
+import '../sections/dropdown_selectors_section.dart';
 import '../sections/shows_section.dart';
 import '../sections/simple_widgets_section.dart';
 import '../theme_switcher.dart';
@@ -43,6 +44,8 @@ class ShowcaseScreen extends StatelessWidget {
                         crossAxisAlignment: CrossAxisAlignment.start,
                         children: [
                           SimpleWidgetsSection(),
+                          SizedBox(height: 16),
+                          DropdownSelectorsSection(),
                           SizedBox(height: 16),
                           ShowsSection(),
                           SizedBox(height: 32),